Add options for serializer to accept a filepath or buffer.
The serializer now accepts an argument specifying a filepath
or buffer to which the serialized object may be directly
written. If this argument is not provided, a JSON string is
returned.

The deserializer accepts a JSON string as before, or a filepath
or buffer from which the JSON string can be read.

Add tests for this functionality, and documentation.

Note: Some of this code was taken from pandas v1.0.1. See
source for attribution.

+++ acnportal/acnsim/network/current.py
@@ -26,7 +26,9 @@ class Current(pd.Series):
         elif isinstance(loads, pd.Series):
             super().__init__(loads)
         elif loads is None:
-            super().__init__()
+            # The object type is specified explicitly here to address a
+            # warning in pandas 1.x.
+            super().__init__(dtype='float64')
         else:
             raise TypeError("Variable loads should be of type dict, str, pd.Series, or Lst[str].")
